+++ protected/humhub/modules/like/Events.php
@@ -29,7 +29,7 @@ class Events extends \yii\base\BaseObject
      * @throws \Throwable
      * @throws \yii\db\StaleObjectException
      */
-    public static function onUserDelete($event)
+    public static function onUserDelete($event): bool
     {
         foreach (Like::findAll(['created_by' => $event->sender->id]) as $like) {
             /** @var Like $like */
@@ -47,7 +47,7 @@ class Events extends \yii\base\BaseObject
      * @throws \Throwable
      * @throws \yii\db\StaleObjectException
      */
-    public static function onActiveRecordDelete($event)
+    public static function onActiveRecordDelete($event): bool
     {
         /** @var ActiveRecord $record */
         $record = $event->sender;
@@ -71,16 +71,12 @@ class Events extends \yii\base\BaseObject
         $integrityController->showTestHeadline("Like (" . Like::find()->count() . " entries)");
 
         foreach (Like::find()->each() as $like) {
-            if ($like->source === null) {
-                if ($integrityController->showFix("Deleting like id " . $like->id . " without existing target!")) {
-                    $like->delete();
-                }
+            if ($like->source === null && $integrityController->showFix("Deleting like id " . $like->id . " without existing target!")) {
+                $like->delete();
             }
             // User exists
-            if ($like->user === null) {
-                if ($integrityController->showFix("Deleting like id " . $like->id . " without existing user!")) {
-                    $like->delete();
-                }
+            if ($like->user === null && $integrityController->showFix("Deleting like id " . $like->id . " without existing user!")) {
+                $like->delete();
             }
         }
     }
